# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Comix manga browsing API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MangaSummary {
  id: string;
  title: string;
  thumbnail: string;
  type?: string;
  isNsfw?: boolean;
}

export interface MangaListResponse {
  items: MangaSummary[];
  hasNextPage: boolean;
  page: number;
}

export interface MangaDetail {
  id: string;
  title: string;
  author?: string;
  artist?: string;
  synopsis?: string;
  altTitles?: string[];
  status: string;
  type?: string;
  isNsfw?: boolean;
  rating?: number;
  thumbnail: string;
  genres: string[];
  score?: string;
  scorePosition?: string;
}

export interface Chapter {
  id: number;
  number: number;
  title: string;
  scanlator: string;
  date: number;
  votes?: number;
  isOfficial?: boolean;
}

export interface ChapterList {
  items: Chapter[];
}

export type PageListPagesItem = {
  index: number;
  url: string;
};

export interface PageList {
  chapterId: number;
  pages: PageListPagesItem[];
}

export type GetPopularParams = {
  page?: number;
  nsfw?: boolean;
  poster?: GetPopularPoster;
};

export type GetPopularPoster =
  (typeof GetPopularPoster)[keyof typeof GetPopularPoster];

export const GetPopularPoster = {
  small: "small",
  medium: "medium",
  large: "large",
} as const;

export type GetLatestParams = {
  page?: number;
  nsfw?: boolean;
  poster?: GetLatestPoster;
};

export type GetLatestPoster =
  (typeof GetLatestPoster)[keyof typeof GetLatestPoster];

export const GetLatestPoster = {
  small: "small",
  medium: "medium",
  large: "large",
} as const;

export type SearchMangaParams = {
  query: string;
  page?: number;
  nsfw?: boolean;
  poster?: SearchMangaPoster;
};

export type SearchMangaPoster =
  (typeof SearchMangaPoster)[keyof typeof SearchMangaPoster];

export const SearchMangaPoster = {
  small: "small",
  medium: "medium",
  large: "large",
} as const;

export type GetMangaDetailsParams = {
  poster?: GetMangaDetailsPoster;
  alt?: boolean;
  score?: GetMangaDetailsScore;
};

export type GetMangaDetailsPoster =
  (typeof GetMangaDetailsPoster)[keyof typeof GetMangaDetailsPoster];

export const GetMangaDetailsPoster = {
  small: "small",
  medium: "medium",
  large: "large",
} as const;

export type GetMangaDetailsScore =
  (typeof GetMangaDetailsScore)[keyof typeof GetMangaDetailsScore];

export const GetMangaDetailsScore = {
  top: "top",
  bottom: "bottom",
  none: "none",
} as const;

export type GetChaptersParams = {
  dedupe?: boolean;
};