# lib/api-zod/src/generated/types/mangaSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Comix manga browsing API
 * OpenAPI spec version: 0.1.0
 */

export interface MangaSummary {
  id: string;
  title: string;
  thumbnail: string;
  type?: string;
  isNsfw?: boolean;
}